💥 refactor(BREAKING)!: switch to `@clack/prompts` for all prompts
This is a breaking change, not because any functionality has changed, because all the core logic is the exact same, but moreso because it overhauls the user experience.

----------------

Along with this, other changes included in this commit are:

- 🐛 fix: description not being filled with user input

- 🧹 chore: use `.name` for showing name of licenses

- 🧹 chore: use significantly better error messages

- 🧹 chore: stop logging when `npm init -y` is ran

I've also removed some extra, unnecessary things like spinners and other things alike.

Something could (hopefully not) break with this change. I've tested it a lot of times, but if you encounter an error, provide the console log, and create an issue.