✨ feat: allow already existing files to be skipped
When you are warned of files being overwritten, you are asked if you want to skip the listed file paths.

If you select 'Yes', the listed file paths won't be created, and the file paths that don't already exist will be created as normal.

If you select 'No', the file paths are overwritten as normal with template content.